+++ src/main/java/subway/line/controller/LineController.java
@@ -1,7 +1,11 @@
-package subway.line;
+package subway.line.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.line.dto.LineRequest;
+import subway.line.dto.LineResponse;
+import subway.line.service.LineService;
+import subway.line.dto.LineUpdateRequest;
 
 import java.net.URI;
 import java.util.List;
