+++ src/main/java/subway/line/entity/Line.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.entity;
 
 import subway.station.Station;
