+++ src/main/java/subway/line/service/LineService.java
@@ -1,7 +1,12 @@
-package subway.line;
+package subway.line.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.line.dto.LineRequest;
+import subway.line.dto.LineResponse;
+import subway.line.dto.LineUpdateRequest;
+import subway.line.entity.Line;
+import subway.line.repository.LineRepository;
 import subway.station.Station;
 import subway.station.StationRepository;
